refactor mask
