+++ llms/mlx_lm/models/llama.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, Tuple, Union
 import mlx.core as mx
 import mlx.nn as nn
 
-from .base import BaseModelArgs
+from .base import BaseModelArgs, create_additive_causal_mask
 
 
 @dataclass
@@ -142,13 +142,6 @@ class TransformerBlock(nn.Module):
         r = self.mlp(self.post_attention_layernorm(h))
         out = h + r
         return out
-
-
-def create_additive_causal_mask(N: int, offset: int = 0):
-    rinds = mx.arange(offset + N)
-    linds = mx.arange(offset, offset + N) if offset else rinds
-    mask = linds[:, None] < rinds[None]
-    return mask * -1e9
 
 
 class LlamaModel(nn.Module):
